use a faster implementation of path.Walk