@komuW 2017-10-20
Closing since there doesn't appear to be much signficant improvement: 


`go test -run=XXXX -bench=BenchmarkBuildDockerImage ./...`          

new(using https://github.com/MichaelTJones/walk):                  
8970 ns/op, 8740 ns/op, 8843 ns/op

old(using stdlib):                
8645 ns/op, 9263 ns/op, 8896 ns/op